feat: Set baseUrl to be retrieved from static path
- Added .gitignore entry to exclude public/config.json
- Set up baseUrl loading from static path
- Implemented get and set methods for baseURL in api.ts
- Updated AppLayout.tsx, apiClient.ts, apiTypes.ts, and main.tsx for baseURL integration
