+++ src/AppLayout.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import RoutingModal from '@/common/components/RoutingModal';
 import { Outlet } from 'react-router-dom';
-
 const AppLayout = () => {
   return (
     <>
