+++ src/main.tsx
@@ -1,9 +1,15 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
+import { loadConfig } from '@/config/urlConfig';
+import { setBaseUrl } from '@/common/service/api';
 
-createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-);
+loadConfig().then(config => {
+  setBaseUrl(config.baseUrl);
+
+  createRoot(document.getElementById('root')!).render(
+    <StrictMode>
+      <App />
+    </StrictMode>,
+  );
+});
